Avoid using message-based IO in the `plaintext` protocol
Can reuse the machinery in `secio` but need to generalize the
"buffering" there
